Add cluster resource discovery helper
