+++ test/helpers/mock_resource.rb
@@ -1,6 +1,9 @@
 # frozen_string_literal: true
 
 MockResource = Struct.new(:id, :hits_to_complete, :status) do
+  self::SYNC_DEPENDENCIES = []
+  self::SENSITIVE_TEMPLATE_CONTENT = false
+
   def debug_message(*)
     @debug_message
   end
